+++ unit/message_unit.cpp
@@ -1,7 +1,6 @@
 #include <vector>
 #include <string>
 #include <xenon/xenon.h>
-#include <xenon/for_each_path.h>
 #include "message_unit.h"
 
 namespace xn = xenon;
